mcp/pyproject.toml
[project]
name = "samvil-mcp"
version = "1.0.0"
description = "SAMVIL MCP Server — session persistence, evaluation history, and evolve convergence for the AI vibe-coding harness."
requires-python = ">=3.12"
authors = [{ name = "insam" }]
license = { text = "UNLICENSED" }
dependencies = [
    "mcp[cli]>=1.0.0",
    "aiosqlite>=0.20",
    "PyYAML>=6.0",
]

[project.scripts]
samvil-mcp = "samvil_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
